+++ common/constants/ipcevents.js
@@ -8,6 +8,7 @@ export const RUN_SCRIPT         = "bd-run-script";
 export const NAVIGATE           = "bd-did-navigate-in-page";
 export const OPEN_DEVTOOLS      = "bd-open-devtools";
 export const CLOSE_DEVTOOLS     = "bd-close-devtools";
+export const TOGGLE_DEVTOOLS    = "bd-toggle-devtools";
 export const OPEN_WINDOW        = "bd-open-window";
 export const INSPECT_ELEMENT    = "bd-inspect-element";
 export const MINIMUM_SIZE       = "bd-minimum-size";